+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Suppression de la fonction WorkerSQL::getRunningWorkerInfo remplacé par WorkerSQL::getRunningWorker #2237
 - Suppression de la fonction WorkerSQL::getJobListWithWorker remplacé par JobQueueSQL::getFilteredJobList #2233
 - Suppression de la fonction WorkerSQL::getNbJob remplacé par JobQueueSQL::getNbJob #2239
+- Le connecteur `pdf-relance` est supprimé. Il est intégré au connecteur `mailsec` #2269
 
 ## Suppression d'éléments dépréciés